Add a linter for dependencies
It's very easy to accidentally #include a header but forget to add the
appropriate dependency to the target. For distro-installed software,
this usually goes undetected by the CI since it all just works, and this
will only get worse with https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/812.

This commit adds a potential solution: introspect the Meson build and
just determine if any headers are used without the appropriate
dependencies. The mapping from headers to dependencies is configurable
but primarily based on distro packages, as reported by `dpkg` and `rpm`.

The CI runs the linter after almost every build, thus ensuring that
every reasonable build configuration properly passes the bar implemented
here.