Correct initialization of static atomic variables
Early C11 required the use of an `ATOMIC_VAR_INIT` macro when writing
the constant initializer for a static atomic variable. The final C11
standard however allowed just writing the value without such a macro,
and the macro was deprecated in C17 and then removed in C23.

Before that, we used to use a hacked-together header in which atomics
were a `struct` type. This meant curly braces in the initializer, but
not all C compilers are happy with initializing a scalar with a
curly-braced expression (invalid in C, but valid in C++).

Remove these ancient support quirks and just initialize atomic variables
the ordinary, standards-compliant way.