Correct fragile MPI sanity check
The Meson build runs a sanity compiler check when using `decode-mpi`.
The C++ code tested refers to the `NULL` constant, but did not include
`<stddef.h>` and so would fail if `mpi.h` did not include it itself.

Fix the check to use `nullptr` instead, which is a C++ keyword-constant
that implicitly converts to `NULL` pointers.