hpcprof-mpi: Correct handling of large counts
MPI functions use `int` as the size (`count`) of the buffers, but
internally we use the more convenient `size_t` instead. Unfortunately
the two have very different sizes on 64-bit machines, the latter being
twice the bytes of the former. When the `count` is very large this
causes MPI correctness errors where we end up passing a negative `count`.

Break down such calls into smaller calls when such a thing is possible,
and add assertions for the others.