Redesign dependency logic for Level Zero
The premise here is to always use the same "method" for finding the
dependency as the upstream project supports. This policy reduces the
number of special cases we need to explain to users before building,
making the build simpler overall.

In line with that premise, we now use the pkg-config file that is
shipped with Level Zero instead of using CMake. We programmatically
work around the bug the files have (https://github.com/oneapi-src/level-zero/issues/131)
using known variables and configuration. It's not a perfect solution,
but we can't do better until Level Zero fixes the problem upstream.

The fallback now uses a compiler check instead of relying on CMake. This
is only necessary for certain incomplete deployments seen in the wild,
and for those it works.