Replace CMake fallback for IGC with compiler check
There are cases where the CMake fallback doesn't actually work, so
replace it with a fallback based on a compiler check.

In proper cases we use a trustworthy pkg-config file, this fallback is
only for certain incomplete deployments seen in the wild. As annoying as
this workaround is, it works in the cases it needs to.