Update README to recommend new wraps
The new wraps make it possible to build with truly zero dependencies,
and they're far more convenient and automatic than trying to use Spack
for this purpose. We may find issues down the road... but so far, it
just works. Update the README with the vastly simplified instructions.

The Spack method of supplying dependencies is no longer required or even
mentioned, there are currently too many caveats with Spack environments
to make supporting it explicitly worthwhile at this time. It works for
applications, it doesn't work for development environments.

Meson wraps are the recommended and supported method, devcontainers are
now documented and supported. We may investigate tools such as Pixi
in the future.