ci: Use OS packages instead of Spack
Instead of installing dependencies through Spack (and then fiddling with
prefixes etc. to make it all work right), just install the packages
straight from the distro. After all, most of our dependencies are
packaged everywhere and easily available. The end result is smaller
predeps images, testing of environments more like real machines, and a
CI with far less loose screws.

For cases where the distro package is too old (below MRV), we let Meson
use the corresponding wrap instead. This mostly affects the oldest
distro in the set, Ubuntu 20.04. We already build Dyninst and Xed from
wraps, so adding Elfutils, Libunwind and XXHash to that list isn't a
huge burden in practice.

The exceptions are Libmonitor and Boost. Libmonitor is built from
source in every single image, Boost is built from source for the CUDA
images. This matches the prerequisites as described in the README.

Also, clean up the structure of the Containerfiles. The base images
don't in general set variables we need to clear, and the set of packages
to install should be a simple sorted list. Git LFS and `eatmydata` have
been dropped as well, we haven't used them for a while. Finally, the
"bare" image no longer has Git, since Git is not a requirement to build.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/812