hpcrun: Only compile PAPI support when PAPI is found
Previously, if PAPI was disabled but the PAPI header was available, a
header check would pass and the PAPI Intel component support would be
enabled... even though the supporting code doesn't get compiled in.
This may have been an issue in the original Autotools, I'm not sure.

This code needs to be carefully disentangled, but dodge this issue for
now with a trivial hotfix.