ci: Clean up and remove extra build configuration
Currently the predeps images contain an `image.ini` native file, which
defines a few paths and properties for Meson so it can find all the
dependencies. This isn't exactly what we want to test, we want to ensure
the tests pass *without* the need for additional configuration.

This commit refactors how the CI goes about its testing to cover this
case better: instead of setting a bunch of feature flags, the option and
semifresh jobs now test with just the flags to be tested, all others
default. This reflects the end-user experience better as well as ensures
the tests are well-covered in this case.

The individual and compiler test jobs still set a bunch of feature
flags, but for consistency these flags are now set in the native file
shipped in the image, `hpctookit-ci.ini`. This is done to remove the
significant duplication of flags for each image, now it's just in a file
shipped with the image. These tests ensure the images support the
required features and that said features work with different compilers.

The dependency linter (lint-depends) is no longer a run_target(), it is
now a test in the test suite. The default interpretations of common
packages have been disabled, all those settings now need to come from a
machine file, as such the test is disabled unless a machine file
explicitly enables it. This makes it easier to transition between linted
and unlinted cases in the CI jobs, the linter runs only when
`hpctoolkit-ci.ini` is added.