build: Search harder for a usable Python
On some distros `python3` is still 3.6, which don't support. If our
initial attempt to find a usable Python fails, search every versioned
Python from 3.8 through 3.14 as a fallback.

Note that since we always search `python3` and `python` first, the user
is still easily able to override the Python used with whatever they
wish, so long as it's >=3.8.