Add -Dextended_tests flag for CUDA and HIP tests
The CUDA- and HIP-based tests require additional dependencies, namely
the compilers for these two languages. This commit adds a new feature
flag for these dependencies, `-Dextended_tests`.

In most cases the compilers will be found, they are shipped with the
monitoring libraries after all. This is mostly a change in anticipation
of more complex cases where the test dependencies *won't* be available
so easily.