predeps: Upgrade the Intel image
The unversioned tags haven't been updated for about 2 years now, switch
to the versioned tags so we can actually test something modern. Intel
is also actively deprecating support for Ubuntu 20.04 so bump to the
next LTS, 22.04.

Also, switch to the "Base Toolkit" instead of the runtime image. This
image is a lot bigger but contains the Intel compiler, icx, which we
need to implement SYCL-based tests.