tests: Add a SYCL test for Level Zero
The test is a simple vector add test, just like the others, but written
this time in SYCL.

Note that we currently assume SYCL is always implemented on Level Zero.
In reality there are other SYCL implementations that target OpenCL,
NVIDIA CUDA or AMD ROCm. The tests will likely fail if any of those
implementations are used here. Caveat emptor.