ci: Drop the dependency linter (lint-depends)
There are a few different reasons for this, but in short lint-depends
has too high a maintenance cost for what benefit it can provide. It has
proven itself by finding current issues but it's just not realistic to
keep it up as dependencies shift.

There isn't another tool that does what lint-depends does *exactly*, but
similar issues should become visible if we compile everything with
`-Wl,--no-undefined` ([#796]) because Meson dependencies represent both
headers and libraries. Note that `--no-undefined` is the default for
Meson, but we explicitly use `shared_module()` and `.partial_dependency()`
for hpcrun to avoid linking certain libraries.

[#796]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/796