ci: Improve spack install job
Instead of only using the buildcache to speed up installing dependencies,
this commit uses the predeps machinery as a first, stronger cache. The
image in this case is designed to be rolling, it uses unversioned tags
for Spack and dynamically adapts the buildcache in use roughly according
to Spack's SemVer version.

Since we're using a cached and derived image now, the `spack install`
job now *always* tests the latest off Spack develop. This means recipe
updates will cycle faster, but there may be instability that appears
when Spack breaks. Caveat emptor, failures are allowed in this job for
this very reason.

The job now also enables both Spack testing options (build-time and
standalone). The test results from standalone testing are reported to
GitLab via JUnit, they aren't required to pass for the reasons above.