Don't enable OpenMP unless necessary
OpenMP pulls in additional libraries, we should be a little more careful
linking it into everything. Adjust the build system so only hpcstruct
and hpcprof have OpenMP enabled, all other tools have it disabled.

This is needed to fix hpcrun when built with the Intel compilers, which
link in libiomp which has init constructors that crash the auditor.

A few other library dependencies are tightened up by this commit, mainly
pthreads, libm and libdl.