hpcrun: Complete conversion of OpenCL support to foils
The main goal of the foils scheme is to allow libhpcrun and the foils
to live in separate link namespaces, thus any code in libhpcrun that
refers to an external library should not use dlsym and instead pass
the functions through the foil->base call.

This commit refactors the OpenCL support code to fully rely the foils
scheme for accessing the target library. Cases where a loose cl*
function was directly called have been replaced by function pointers,
cases where the symbols were fetched through a manual dlsym() have been
rerouted to come from the foil interface instead. opencl_bind() has been
removed as a result.

A number of the internal interfaces have changed as a result, but the
only logic change is that the total number of available devices is now
counted during clCreateContext instead of end-of-program. The overall
result is not expected to change because of this.