tests/hpcrun: Test different threading models
There are different models for how threads can be created and do work.
It's worth adding tests for a few more than just OpenMP, so add explicit
pthreads, C++ std::thread and C11 threads to the test matrix.

Note that C11 threads are XFAIL because libmonitor doesn't intercept
the functions for that model. Caveat emptor.