hpcrun: Use a special exit code for some CUPTI errors
CUPTI requires additional setup on the part of the sysadmin to function.
The profiling hardware is also a highly limited resource and we may not
be able to get it. To make this work in automated setting (e.g. the test
suite) give these errors a special exit code.

The exit code used in this case is configurable. It defaults to 1, the
generic error code, but the test suite uses 77 to cause test SKIP.
Available as a control knob: `-ck HPCRUN_CUPTI_SYSTEM_ERROR_EXITCODE`