Compile for all available CUDA arches
When compiling CUDA code, we want to support as many possible base CUDA
architectures as possible. This commit adds some logic so that we always
compile for all the CUDA architectures the running NVCC can support.

Right now this only affects the tests, we don't compile any CUDA kernels
anywhere else. But if we ever did, we would want this too, so apply the
flags generally to the whole project.