hpcrun: Hotfix the control knobs subsystem
The way the logic works, if a control knob is registered after the call
to control_knob_init(), the default ends up overwriting whatever the
user entered on the command line. But since control_knob_init() is
called right at the start of execution, control_knob_register() can't
be called before it by the sample sources etc.
This is very obviously not correct.

The control knobs interface could do with a complete interface redesign,
but for now make it work by making control_knob_register() not overwrite
the value if it was already set by the user.