hpcrun: Complete conversion of libc overrides to foils
The main goal of the foils scheme is to allow libhpcrun and the foils to
live in separate link namespaces. This means when libhpcrun goes to call
the "real" symbol for an interception, it needs to get the right
"instance" of the symbol, specifically the one the foil observes.

Previously the interception code in libhpcrun for the pthread-blame,
and io foils would call an underscore-qualified symbol defined by GNU
libc. This commit refactors these foils to pass through the instance
from the foil's perspective, using dlsym() like all the others.

The interception code for memleak foils uses the understore-qualified
name to avoid recursion problems with Glibc.