hpcrun: Convert the launch script to C++
As our initialization code gets more complex, we need earlier and
earlier "hooks" to attach to with `gdb`. As it stands right now, you
can't just `gdb --args hpcrun -e REALTIME ./my_app`, since `hpcrun` is
a shell launch script and `gdb` only operates on (binary) executables.
We have the `hpcrun -d` flag for this case which invokes a spin-wait
for attachment, but for some initialization tasks this is still "too
late" of an attachment point to observe everything going on.

Shell is also notorious for its many flaws (non-portability, inelegant
and insufficient error handling, lack of first-class values, etc.) that
make it unsuitable for large and robust applications. The `hpcrun`
launch script is >850 lines of complex logic. Times need to change.

This commit converts the script to C++, as close to 1-to-1 as possible.
A few difficult logic pieces that probe the application binary have been
removed, most of these only generated early errors and so have little
effect on users. The logic to find libmonitor has also been rewritten.

And yes, `gdb --args hpcrun ...` just works now. You're welcome.

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/699 and https://gitlab.com/hpctoolkit/hpctoolkit/-/work_items/703