devcontainers: Integrate with Clangd extension
Clangd requires a reference to the build directory to properly find the
compile_commands.json. In devcontainers the expectation is that the
Meson extension is being used, so add (default) configuration for the
Clangd extension will use the Meson extension's build directory.

Note that this doesn't add Clangd as an extension, it only adds the
configuration to make it work.