Fix segfault on OpenCL test skip
If GPU monitoring was enabled but the application did not actually
initiate any GPU operations (e.g. no GPU available), then the channel
would not be allocated and thus shutting it down would segfault.
Avoid this by ensuring the channel is fully initiated before attempting
to tear it down, possibly allocating it if it wasn't already.