hpcprof: Fix rare crash in SparseDB
Sometimes a whole bunch of the tests fail due to a crash in SparseDB,
where a std::vector used as a heap ends up empty during an output loop.
Fix this by checking whether it's empty beforehand.

It's a little unclear why this happens inconsistently, it may be related
to the filesystem iteration order or may be something completely
different. Regardless, this fixes things.