hpcprof: Remove one-after-end array/vector access
There are a few places where the one-after-end element of a std::vector
or std::array is accessed or referred to. Newer(?) versions of the STL
consider these to be assertion-worthy errors, so rewrite the code to
avoid touching said elements.

(cherry picked from commit 3053d14145487fdec41776a2d96a77b32e42bcc8)