Use proper Python for decode-mpi
The Python scripts in the repository are written in Python 3.8 and have
proper shebangs, however Meson 1.4 changed how find_program() behaves
with such scripts and as a result no longer pulls the right Python.
This only affects decode-mpi, so change the call to use the Python
found earlier via an explicit find_program() call.