hpctesttool: Allow "0" in address ranges
Previously the regex would only accept address ranges if both endpoints
were in explicit hexadecimal, i.e. `0x[0-9a-f]+`. Sometimes though 0 may
be represented as just `0` instead of `0x0`. Accept this.

The cases where this happens contain other bugs, so it's unclear if this
should or shouldn't be accepted, but this whole format needs redone
anyway so it's fine I guess.