Remove #if 0 "comments"
One *common* way to comment out a block of code is to wrap it in #if 0,
which commands the preprocessor to elide everything that happens until
the following #endif (...kind of, there's some catches there regarding
nesting). These code segments can appear anywhere, and usually contain
code that was thought about and retained because "it might be useful
someday." Most outlived their usefulness many years ago, and many are
bitrotten past being usable at all.

The end result is a waste of disk space, CPU cycles, and eyesores for
anyone trying to work on the codebase.

This commit elides them all, with absolutely no regard to authorship or
purpose. The effective code has not changed a byte.