hpcrun: Don't predefine _GNU_SOURCE etc.
Previously we would predefine _GNU_SOURCE (and __HIP_PLATFORM_*) in all
hpcrun source files. This is inconsistent with the rest of the codebase
and isn't the recommended way to use these macro flags.

The Meson build no longer injects -D flags for these macros, the source
files now #define them at the top. This is almost certainly overkill and
in many cases could be removed, but this way is easier to get things
to compile first.