hpcrun: Promote custom -D flags to project
Currently there are a number of -D flags passed only to the hpcrun code.
Even though these flags only affect the hpcrun code, this creates an
inconsistent source of truth between the hpctoolkit-config.h and the
build environment.

This commit promotes these flags to the project level, making the build
environment consistent across all source files in the repository.