Clean up a few improper preprocessor usages
There are a few source files that #ifdef the entire file as a way to
ensure it only gets compiled when certain features are enabled. This is
silly, so this commit converts them to use the build system instead.

Also, there are a few (normal) headers that don't have proper header
guards. This commit adds/fixes those too.