Move --version features message to meson.build
Previously we would set #defines for the various versions and features
that were enabled. This works, but causes warnings and isn't a
particularly clean approach.

With this commit, the logic for the features message is now implemented
in the meson.build, which has full access to the dependency() objects
and thus doesn't need as much complication. The string is exposed via
hpctoolkit-version.h as a new printing function. (Due to the use of a
raw string literal here, this is only available for C++ code.)