Drop Spack spec message from --version
This hasn't worked at all since the Meson rewrite, so just drop the
line in the message. Helps make things work better.

This is also the last "special" thing we do for Spack. Spack is just
another deployment platform, there really is no reason for us to cater
to it specially. And being real for a second, most of the problems we
see in Spack builds are Spack's fault and not ours.