Add OpenMP placeholders so that hpcreport can distinguish parallel regions and tasks
This commit was squashed from the following commits:
* fixup! hpcrun: add <ompt work> placeholder

Disable OMPT_DEBUG

* fixup! hpcrun: add ompt placeholders for impl and expl tasks

Remove now-invalid comment

* hpcrun: avoid empty inner frame

removed blank innermost frame added as padding to avoid a
buffer overrun in backtrace buffer. the padding was visible
and causing an empty inner frame.

the need for the padding is no longer present; remove the
padding to finish the cleanup.

* hpcrun: correct unused variable

* hpcrun: fix indenting in ompt-callstack.c

* hpcrun: add ompt placeholders for impl and expl tasks

* hpcrun: add <ompt work> placeholder