Generate sources not headers
There are a few places where we generate headers, for instance
`src/lib/xml/hpc-structure.dtd.h`. Convert these to generate sources
instead, which get compiled into applicable build targets.

There are two advantages here. One is that the interface header is at a
constant location in the source directory, which requires less
`include_dirs` to permit inclusion at than a header in the build dir.
The other is that changes to the generated output build much faster,
since the header is not generated only the (relatively small) generated
source needs to be built and the target relinked.