Share sources instead of static_library
Previously when code had to be shared between tools, it lived in
`src/lib` and was compiled as a shared_library linked into the tools.
While this works, it also leads to complications as there are a number
of build_target() arguments that need to be consistent across the link.

Avoid the hiccups by sharing the source files instead of the compiled
results. This means some source files will be compiled multiple times
as part of the build, but this allows the arguments for each target to
differ and so be tuned for it's specific purpose.