hpcprof: Enable DirectClassification under --foreign
We've always disabled DirectClassification (DWARF/ELF parsing) in the
--foreign case, since DirectClassification is about analyzing the binary
itself and --foreign is about *not* doing that. But there actually a few
cases where the binary may be transferred: gpubins and vdso come as part
of the measurement directory.

To make this all work right, we refine how `resolvedPath` is understood,
it is simply the path "on the current filesystem" and may not actually
exist in practice. The final behavior doesn't change (although a
conditional was moved from DirectClassification to Pipeline).

The difficulty here is that there is no good way to tell if a Module
represents an actual binary on the filesystem. Currently that is assumed
to be `.path().is_absolute()` (i.e. "real" binaries are absolute paths).
It works well enough for the way hpcrun operates today.