tests/hpcprof: Pass -R flag for measurement dir
Measurement directories currently can't be renamed, it is a known issue
that hpcrun produces absolute paths to the measurement directory in
many circumstances. The solution for now is the oft-forgotten
`-R from=to` flag, which remaps a tree `from` one path `to` another.
Pass this flag in the tests to remap the measurement directory to its
current location.