ci: Test case where gettext requires libintl
The *gettext() family of functions are for querying translation
databases. They aren't standard but are provided by Glibc, however it is
possible to use an external implementation instead, GNU gettext. This
implementation requires linking against libintl to provide the bits, so
there may be an extra step to the build.

Add a CI test for this very scenario, using an image based on the `bare`
but with a custom build of gettext forced onto the compiler.