ci: Include test error logs in CI job log
Currently, the only way to see the error from a failing test is to check
the JUnit output as parsed by GitLab. This is a perfectly fine interface
for deep inspection of a test log, however the UI workflow is not great
when the test failure is possibly expected and a cursory glance is all
that is required.

Assist the cursory inspection by passing `meson test --print-errorlogs`,
which includes the first 100 lines of any test failures in the job log
of the failing CI job itself.