tests/data: Don't apply meas dir remapping on regen
Follow-up to https://gitlab.com/hpctoolkit/hpctoolkit/-/merge_requests/1096.
That MR added an option to remap the measurement directory when running
hpcprof tests, however when test data is regenerated the same remapping
is applied, so the regenerated data will be looking in the previously
committed measurement directory instead of the newly generated one.
This causes inconsistent results when regenerating test data and often
causes the semi-fresh jobs to fail in CI.

Sort this out by omitting the remapping flag when regenerating the data.