Make tools and common code siblings
Until now, all shared code had to be in src/lib while the tools lived in
src/tool. This directory arrangement was necessary to ensure proper
build order when using Autotools (possibly only for older versions).
On Meson this is no longer an issue as directory distinctions do not
affect the build order, so we can do a mass rearrangement of the code to
what it really should be.

The tools are now in `src/hpc*`, one directory per tool. The common
code is under `src/common`. The few vendored headers from Valgrind are
under `src/extern` as before.