Merge shared code under src/common/
We've had distinctions between the common code pieces that really just
doesn't mean much in the current world. Remove (most) of the
distinctions and merge everything under src/common/.

One distinction is still made here, code that is pure-C and thus safe to
use in hpcrun (i.e. prof-lean/, support-lean/, most of include/) is now
src/common/lean/, while the remainder is in the general src/common/.
Eventually this untoward distinction will also be removed.

Note that distinctions are removed in the build system as well, tools
now build against `common[_lean]_srcs` and `common[_lean]_deps`.
Generally this means more bits are linked into each executable and there
are more duplicate builds, but the effect is expected to be minimal.