Merge lib/ into src/ and doc/
lib/ doesn't have a clear role in the top-level organization of the
project... so remove it in favor of those that do.

`lib/python` is now `src/hpctesttool`, following the pattern of tool
sources being under `src/`. Note that this is still a Python tool, so
the internal structure is completely different and even reads a little
funny (`src/hpctesttool/hpctesttool`), but this is correct.

`lib/dtd` is now `doc/dtd`, as the DTD is at best just documentation, if
even that. It's basically bitrot at this point, but until it's removed
we move it here.