hpcrun: Don't intercept dlmopen() in fake-auditor mode
hpcrun has two "modes" for tracking libraries: an LD_AUDIT (auditor)
mode and an LD_PRELOAD (fake-auditor) mode. The latter intercepts dl*()
functions to track when the application dynamically loads libraries.
Unfortunately, this interception causes said dl*() functions to resolve
against the RUNPATH (or RPATH) of `hpcrun` itself rather than the app,
which can cause application failure. As a rule, the auditor mode
performs better in such circumstances and works better in these cases.

The auditor mode has issues of its own, but a major one is that many
versions of Glibc crash when dlmopen() is used to create a namespace.
GTPin is one such library that does this, and we use it sometimes.
This encourages the use of the fake-auditor mode (even by default) in
certain cases, however since dlmopen() is intercepted it runs into
R*PATH issues. GTPin is one such library that is affected by this.

One solution to this catch-22 is to not intercept dlmopen() in the
fake-auditor mode, this commit implements this solution. This makes
logical sense: the fake-auditor mode exists primarily to work around
bugs that exist in many Glibc versions, and a major bug is dlmopen(),
therefore the workaround should have no caveats for this use case.
This also happens to match up with `hpcrun`'s capabilities pre-auditor.

An XFAILing GTPin test is fixed by this change.