Convert to REUSE licensing metadata (part 1)
License headers are long and hard to read for very little gain overall.
It's also very easy to accidentally mangle the license header, or cause
confusion when multiple copyrights/licenses are at play. Convert the
codebase to consistently use the [REUSE Specification] to express the
licensing and copyright metadata in a human- and machine-readable form.

In short, this boils down to a comment at the top of every file listing
one (or more) SPDX-FileCopyrightText and SPDX-License-Identifier
attributes. These few lines apply to the file that contains them, which
allows per-file licensing among other perks.

This commit is the easy cases, files that already have (our) copyright
headers and just needs some quick translation to match with REUSE.
"Unimportant" files that only contained the Rice copyright are made
completely empty to avoid unnecessary confusion.

[REUSE Specification]: https://reuse.software