hpcrun: Support immediate command lists with GTPin
The GTPin code assumes the GTPin callback comes in before the command
node is processed at the Level Zero layer. That doesn't happen when the
CommandList is immediate, the Level Zero command node is applied
immediately before the real Level Zero function is called (which causes
the GTPin callback to fire).

Reorder the node processing until after the real Level Zero call,
restoring the order the GTPin support is expecting in this case.