Convert to REUSE licensing metadata (part 2)
License headers are long and hard to read for very little gain overall.
It's also very easy to accidentally mangle the license header, or cause
confusion when multiple copyrights/licenses are at play. Convert the
codebase to consistently use the [REUSE Specification] to express the
licensing and copyright metadata in a human- and machine-readable form.

This commit handles the cases that have foreign copyrights, for instance
the fragments of the Intel GTPin examples or the Valgrind vendored
headers. The "Copyright" lines are adjusted to REUSE machine-readable
format (`SPDX-FileCopyrightText:`) and `SPDX-License-Identifier` is
added just after the corresponding license text.

Note that, annoyingly, not all the licenses identified are perfectly
compatible with the BSD-3-Clause. In practice there's ~5 files that have
additional restrictions beyond what BSD-3-Clause enforces, but bleh.

[REUSE Specification]: https://reuse.software