Convert to REUSE licensing metadata (part 3)
License headers are long and hard to read for very little gain overall.
It's also very easy to accidentally mangle the license header, or cause
confusion when multiple copyrights/licenses are at play. Convert the
codebase to consistently use the [REUSE Specification] to express the
licensing and copyright metadata in a human- and machine-readable form.

This commit handles our files that, for one reason or another, were
missing copyright/license blocks. The years listed in the copyrights are
derived from the Git history of the individual files.

[REUSE Specification]: https://reuse.software