Convert to REUSE licensing metadata (part 4)
License headers are long and hard to read for very little gain overall.
It's also very easy to accidentally mangle the license header, or cause
confusion when multiple copyrights/licenses are at play. Convert the
codebase to consistently use the [REUSE Specification] to express the
licensing and copyright metadata in a human- and machine-readable form.

This commit handles vendored files that are missing copyright metadata.
In some cases this is because the license is contained in a zipped form,
but the other cases are more headache-inducing and a conservative option
is listed for metadata purposes.

[REUSE Specification]: https://reuse.software