Convert to REUSE licensing metadata (part 5)
License headers are long and hard to read for very little gain overall.
It's also very easy to accidentally mangle the license header, or cause
confusion when multiple copyrights/licenses are at play. Convert the
codebase to consistently use the [REUSE Specification] to express the
licensing and copyright metadata in a human- and machine-readable form.

This commit adds the final pieces to achieve and maintain compliance:
the LICENSES/ texts, a CI job, and a bit of extra header text in the
top-level LICENSE file.

Out-of-line *.license files are produced as part of the test data
regeneration process.

[REUSE Specification]: https://reuse.software