ci: Disable jobs in CI runner "unprivileged" cases
As we get external contributors, there will be more than a few cases
where the pipeline is unable to access either the custom runners or
GitLab fleet runners that require a subscription. In these "unprivileged"
scenarios, skip over the problematic jobs and let the CI pipeline run.

An assumption made here is that only merge request pipelines can fall
into this scenario, as running the pipeline with privileges produces a
merged results pipeline (and merge trains should be merged results).