ci: Redesign the Spack install check
We run `spack install` as part of our CI to try and detect issues with
the Spack recipe (stored in the Spack repo [here][1]) before we do
anything rash. The issue is a `spack install` can take quite a long time
installing all the dependencies... and we're impatient. So until now
there's been a bunch of clever container tech to try and speed up Spack
builds, designed to dynamically update as Spack itself evolves.

This cleverness is no longer working. Spack core is a touchy mess when
it comes to buildcaches and reuse, and we're poking directly into the
heart of the problem. Even when it works, it usually rebuilds everything
and takes >50 min. Something has to change.

This commit completely redesigns the job. Instead of container tech we
only use Spack buildcaches, and use a `develop-DATE` snapshot buildcache
instead of the badly pruned `develop` buildcache. Errors that probably
aren't our fault are considered "temporary failures" and ignored.
The end result is a better job that runs faster overall.

[1]: https://github.com/spack/spack/blob/develop/var/spack/repos/builtin/packages/hpctoolkit/package.py