Convert the manual to Sphinx/MyST
This commit was squashed from the following commits:
* manual: Remove the rest of the math

"Math" is a role in Sphinx to include Latex-style
math in documents. Unfortunately it has multiple
consistency and quality issues with renderers for
offline reading. We also don't really need math
text in the vast majority of cases.

Go through and convert the remaining instances to
Markdown-standard inline code blocks (or similar),
and disable the `$...$` syntax for writing math.

* Fix hpcviewer's manual text format

This commit was squashed from the following commits:
* Fix adding the dollar sign

* Add a space before the equal sign

* Agree adding a space

* Update fixes from pre-commit

* Manual: fix hpcviewer section

* Manual: Remove latex symbol in hpcviewer section

* ci: Enable documentation rendering

The new documentation framework requires additional dependencies: Sphinx
and MyST-parser. Add these new dependencies to the images so CI can
operate properly and ensure the documentation builds to completion on
various versions and Pythons of those tools.

PDF output is also supported, but requires an obscene fraction of a
TeXLive installation to operate. So a new "docs" predeps image is added
with these packages, and a new CI job renders said documentation and
exposes it as a browsable artifact for the MR page.

* docs: Render SVGs to PNGs

For HTML documents to behave well on monitors of various scales, the
recommendation is to use scalable graphics (SVGs) for diagrams and
figures, instead of fixed-size images (e.g. PNGs).

Unfortunately, pdflatex doesn't work with SVGs, it only works works with
PDFs. PDFs are also scalable graphics but don't play nice with HTML
web browsers. Let that sink in a moment.

This commit converts the SVGs to PNGs, which are universal to both
formats.

* docs: Make warnings errors, and fix the warnings

At this point the warnings mostly have to do with header tricks Pandoc
tries to use for something like a definition list. They don't work in
MyST because of limitations on headers in rST, so reforge them as actual
Markdown definition lists.

Note that definition lists are a fairly awkward extension to CommonMark,
they aren't supported on GitHub or GitLab or many other places. So we
should *probably* rewrite the text so they aren't needed, but until then
this works well enough.

* Slice up the User's Manual into doc/src/users/

Sphinx uses the file structure (in part) to determine how the
documentation is meant to be arranged. Split the User's manual in the
simplest possible way: by chapter.

Except for the front matter (introduction, quickstart etc.) chapters
with images are contained in a subdirectory. It's a simple organization.

Some figures have been removed in this commit, they don't seem to be
referenced anywhere. They're still recorded in the Git history so there
should be no issue with leaving them behind.

* manual: Last bits of typographic cleanup

Mostly related to code regions, removing leading spaces and adding
language codes where appropriate.

* manual: Strip unused anchors

Pandoc generates explicit anchors for every possible reference point,
but we don't use nearly all of them. Strip out the ones that aren't used
so we can focus on the ones that are.

* manual: Remove invisible and strange Unicode

Pandoc produces Unicode as part of it's Markdown output, with
non-breaking spaces and other little quirks. Unicode is hard to work
with in practice, so this commit removes any Unicode that is invisible
or incorrect (e.g. `--disable-auditor` not `–disable-auditor`).
The remainder are characters like "é" which must be Unicode.

This also enables MyST "replacements" to use readable ASCII
representations of some more common Unicode characters (e.g. the dashes
"–" and "—", and ellipsis "…"), and "smartquotes" for fancy looking
quotes (“”).

* manual: Convert HTML to MyST

Pandoc produces a bunch of HTML code for various structures that aren't
readily available in Pandoc's Markdown. Some (like figures) have proper
syntaxes in MyST's Markdown, others (like "smallcaps") don't make sense
everywhere and so are removed.

This commit goes through the raw Markdown conversion and translates what
can to MyST. There are a number of locations where the translation is
imperfect or just doesn't look right, this is the nature of the problem
of converting between formats.

* manual: Various adjustments to pass pre-commit

This passes the code through mdformat to remove weird spacing errors,
and fixes a few spelling errors picked up by codespell.

* Pandoc the PDF manual to Markdown

See following commits.

This commit is split out to help with rebasing. The changes here were
mechanically produced by executing:

    pandoc -s --wrap=preserve -f latex -C \
      -t commonmark+attributes+smart+tex_math_dollars+strikeout+footnotes+pipe_tables+task_lists \
      HPCToolkit-users-manual.tex -o HPCToolkit-users-manual.md

After making a number of adjustments to the original LaTeX source to
smooth out the translation as much as possible.

The `fig/` subdirectory is also included but (a) PDFs are converted to
SVGs and (b) unused figures are not included.

* Bootstrap a Sphinx manual

Our documentation is in a sorry state. Our user documentation is stored
as an uneditable PDF, ~135 pages generated from a similarly large amount
of LaTeX code. We have no directory to store developer documentation,
the best we have is hand-written text files in no known markup language.
[#716] lists more of the sordid issues.

This commit adds a new documentation framework based on Sphinx, the
documentation itself is written in CommonMark, specifically MyST.

By default static HTML and EPUB documents will be produced, these are
available in `builddir/doc/src/html` and `builddir/doc/src/epub/*.epub`
respectively. This is controlled with the `-Dmanual` feature flag.
PDF output can be enabled via the `-Dmanual_pdf` feature flag, this
additionally requires a Texlive installation for the final render.